fixup: Decreased defines.h footprint
I did not want to move these constants because they are conceptually
wrong, and because one of them should be completely removed along with
deprecating the associated field. However, we can just document the
known problems explicitly and still reduce defines.h footprint.

I did not rename the constants in this commit because they should be
renamed when their definitions are adjusted to address XXXs, to clearly
show the impact of those important fixes. For now, this is just about
defines.h minimization.