Validate --only-changed-since argument before using it
... to be able to provide a more meaningful error message when a
developer makes "fork" typo or copies commit SHA from the wrong repo.

    fatal: Not a valid object name forc
    ERROR: Cannot use --only-changed-since reference point: forc
    Consider using a git commit SHA (from git log) instead.

When "git diff X" fails to find files, we cannot easily tell whether it
failed because of an invalid X or because there were not changes since X
(e.g., when X is HEAD). We could (and probably should!) examine its exit
code, but we do not (XXX?).