fixup: Use a more direct $OnlyChangedSince check
The validity of various git objects is context dependent, so we should
not validate an object used for `git diff` with `git cat-file`. This
change also addresses an XXX in the previous branch commit by making
sure `git diff` and `git ls-files` do not fail.

Also warn if we have no files to scan/format. It could be OK, but it is
usually a problem (e.g., specifying the wrong --only-changed-since SHA
or failing to auto-compute the right forking point).