Compile-time callback type check
The lack of this check was marked as "problem 3" in the previous commit.

We could template the constructor instead, but that would mean inlining
a lot of internal code or forwarding a lot of parameters to another
constructor or method. If the caller forgets to configure the callback,
the bug is likely to surface very soon (unlike some other optional
parameters that are configured post-construction).

At least with HappyConnOpener and PeerConnector, there are already
post-construction job configuration steps, so we are not making things
worse.

The asyncCall() return type improvement was necessary to preserve the
exact AsyncCall type for the Callback::set() to pick up. It will also
remove a lot of explicit AsyncCall::Pointer types throughout the code!