fixup: Removed Ipc::Inquirer callback
It turns out that Ipc::Inquirer code does not use a regular callback
mechanism where code A gives code B an async call C to parameterize and
schedule at some future time, where A is not B and, hence, B does not
know C details. In this case, A is B and knows all C details; there is
no "how to set parameters hidden in some async call given to us" problem
here. In fact, there is no need to create that async call in advance!
The call can (and should) be created (with the right response parameter)
when and where the response becomes known. We now do that.