fixup: Converted Ipc::StartListening() callbacks
Kept ListeningStartedDialer because the current client_side.cc code
relies on delivering additional, non-answer parameters to the callback.
Normally, this additional information is stored inside the job object.
In this case, there is no job dedicated to awaiting a port opening, so
we use a custom dialer to keep that information during the wait.