Stop using raw pointers for jobs with branch-modified constructors
This slightly out-of-scope change does not increase the diff much
because all the affected construction calls need to be modified anyway.
On the other hand, it prevents (unlikely) leaks if the job
configuring/starting code throws. It also prevents requests to prevent
such leaks "since we are modifying this code anyway".

Will move into a dedicated branch/PR if others insist on such a move.