fixup: Removed "wrong correlation"
The earlier change was just trying to get rid of a Gopher reference
(without restructuring the paragraph text which some may view as an
out-of-scope change).

The reviewer appears to dislike the essence of the official
paragraph. I agree that the paragraph is problematic.

Since the ServerProtocol doxygen defgroup and even the concept of
defying such groups are deprecated, it is probably best to just
delete the whole paragraph and move on.

Context:
https://github.com/squid-cache/squid/pull/1092#discussion_r923979458