fixup: Do not call expensive hasReadableDiskEntry() in vain
Once we start swapping out, the (positive) swapOutDecision is final.
There is no need to call expensive hasReadableDiskEntry() in that state.
I bet this is more of a polishing change than optimization because

* before swapout start: The swStarted condition is false so the
  placement of its if-statement is immaterial.

* during swapout: mayStartSwapOut() should not be called at all.

* after swapout ends: I doubt mayStartSwapOut() is called often; there
  should be few trigger events that could tickle this code path.

However, fixing/polishing the comment and addressing TODO is worthwhile.

----

Cherry-picked commit 08fdf89 that (temporary) lives on the
SQUID-762-fewer-findPreviouslyCachedEntry-bag51 branch. I am not 100%
sure that this change belongs to this branch, but this branch does
modify related if statements, so improving their descriptions is
definitely in scope. And once descriptions and understanding of those
statements are improved, reordering them feels acceptable, especially
since the order should not have any significant effect here.