Protect branch-touched swapOut() code from missing Store::Root()
This change is slightly out of scope, but we are adding more
Store::Root() calls and moving old Store::Root() statements around.
Every time we do that, questions about Store::Root() availability are
raised. More shuttting_down checks is not a proper long-term solution,
of course, but it is a simple way to squash those questions.