Removed code duplication that led to a branch bug
... and needlessly exposed Connection users to the COMM_BIND_NOW hack.

COMM_BIND_NOW was missing for one of the two comm_open_listener()s...
Currently, only one comm_open_listener() caller needs that flag, but we
do not maintain and, hence, cannot rely on that invariant. There is also
no good reason to introduce a yet another subtle difference between the
two comm_open_listener() variants.

Yes, we now set COMM_DOBIND twice (unless the compiler is smart enough
to optimize that away). However, the overhead of doing so is small and,
if somebody has the time to check all COMM_DOBIND uses, the first
setting can be eliminated completely if we do not record the flag in
conn->flags (TODO). I could not quickly make sure the old flag is never
needed in Connection outside the comm_open_listener() chain in comm.cc.