Reversed (negative) logic: Enable the optimization on bind+connect path
... instead of disabling it on the bind(2)+listen(2) path.

This allows us to skip all the checks on the bind(2)+listen(2) path that
does not need, does not care about, and can be harmed by this
optimization. The optimization is meant for the bind(2)+connect(2) path!

Also, for many humans, it is easier to deal with a "positive" "do it"
flag than a flag that prevents something from happening, especially when
that something does not happen all the time anyway.