fixup: Fixed the list of Makefile squid-conf-tests dependencies
Duplicating these file patterns is wrong, but I could not find a better
way to do this without moving file filtering into the shell script. I
considered using a "case" statement to filter `*.conf*` files, but
feared that it may not be portable because only one Squid Makefile.am is
using a "case" statement (translate-warn target in errors/Makefile.am).