Polish use of SQUID_YESNO macro
This macro performs a very specific validation check
but was passed a mix of different error messages to
display on failure. Many of which were unclear or wrong.

Move the text of the error inside the macro to enforce
consistency and accuracy about what failed. It actually
only needs to be passed the name of the ./configure
options which was configured badly.

Also, the options configured can only take the positive
form when they fail '--with-foo=X' or 'enable-foo=X'
so update the second argment to always show that form
regardless of whether the configure.ac logic has it
default enabled or disabled.

Also, fixed a bug in --with-nat-devpf validation
which was testing the wrong variable 'enableval'.
This bug only affects the errors displayed so no
builds were significantly affected.