Report (problems with) DEAD-at-start cache_peers
When a cache_peer is created during (re)configuration, Squid probes its
address using a throw-away TCP connection. If and only if that TCP probe
succeeds, peer selection algorithms may pick the peer. Until then, the
cache_peer is treated as if it was DEAD.

This change preserves the logic summarized above but starts _reporting_
the initial probe failure and the fact that it marks the cache_peer
DEAD. Without these reports, admins often naturally assume that the
cache_peer is alive, especially if traffic can be served without it.