Drop LIBCAP_BROKEN workaround hack
The logic setting this #define has been broken since 2010.
The hack was enabled for cases where it was unnecessary,
and disabled when it was apparently needed.

Yet nobody noticed. Implying that the hack is no longer
needed.

For now leave the test for brokenness in place. It
was apparently there for Solaris which we currently
do not have test replication ability for.

Instead of enabling the hack either disable the library
or cleanly abort build if the admin has indicated it
as a requirement.