Redux broken krb5.h handling
De-duplicate the #include sequence into a libcompat header
to ensure all uses of the hack(s) including autoconf
test macros are using the same logic. This fixes some
macros which were using inconsistent logic for Solaris.

Simplify the if-elif use of __cplusplus detection. All
the extern C hacks apply only when C++ is used.

Wrapping the entire file peer_proxy_negotiate_auth.cc
with 'extern C {...}' is unnecessary (it is a .cc)
and breaks the Solaris workaround.