Polish krb5 feature tests which try to compile
Update to AC_LANG_PROGRAM and AC_CACHE where needed.
Break out into separate macros for easier maintenance.

Fix bug where the test for max_skew was never being run.
Replace the #ifdef hack with a libcompat replacement function
as-needed and only test for access to it if a function
is not provided by the library. Also cope with cross-compile properly.