fixup: Remove unnecessary branch-added #includes
We should not rely on CachePeer.h including acl/FilledChecklist.h, but
NoteOutgoingConnection*() function do not really belong to CachePeer.h
in the first place. Ideally, they should be in a dedicated header that
does not expose all CachePeer users to that rather heavy functionality.
However, templates complicate creating such a header.

Perhaps it will be easier to resolve in C++14. And at that time, if this
code uses "auto" instead of ACLFilledChecklist, there will be no need to
explicitly include acl/FilledChecklist.h.