Obey the branch-defined ACLFilledChecklist::setX() API
    FATAL: assertion failed: FilledChecklist.cc:280: "!request"

Security::PeerConnector::sslCrtvdCheckForErrors() sets the request when
creating the ACLFilledChecklist object. NoteOutgoingConnectionFailure()
and similar code paths rely on the third party to create that object, so
they must supply the request themselves to the already created object.
To avoid checking whether the request is set (and hitting assertions
when we forget), it is best to check the request inside setRequest().