fixup: No Foo:: inside Foo
The old Http::StatusCodeString() violates this principle but I do not
think we have to do that. There is other official code that does not use
Http:: for status codes inside the Http namespace, so we are not
introducing a new inconsistency here.

Also polished formatting of this moved code.