fixup: Do not ban retries based on unset HTTP status code
AFAICT, FwdState::reforward() checks IsReforwardableStatus() for
"completed" entries, after the response header is received (or replaced
with a generated error), when the response status code is probably
known. It does not check whether it is known (which is hard to do!).

TunnelStateData::retryOrBail() checks IsReforwardableStatus() in cases
when we have not seen a response yet. We have to check.

XXX: I am not sure how to distinguish zero response status from the
unset one. A _received_ zero status code will probably result in a
parsing exception/error, but I am not sure we will not get to the
IsReforwardableStatus() check with a zero status code in that case. If
we do, we should ban reforwarding rather than skip the check (and allow
reforwding). We should test this.