fixup: Removed oversimplified TODO
I tried to enact this TODO and discovered that peerDigestCreate() stands
in the way. At the minimum, that call would have to be moved lower,
after we register the CachePeer because the PeerDigest constructor locks
the CachePeer object.

Perhaps more importantly, we need to switch from cbdata locks to
refcounted CachePeers because a reconfiguration must not yank CachePeer
objects from under the feet of cache_peer-using transactions!