Do not overwrite caching bans
To ban caching, Squid was setting RequestFlags::cachable to false in
several places while also setting it to true based on a positive
maybeCacheable() decision. That combination could only work as intended
if the ban always came after the maybeCacheable() call, but the code did
not (and should not) enforce such a complicated timing invariant.

Squid now correctly honors caching bans, regardless of the call order.

This change addresses an old XXX, but we are not aware of any specific
bugs fixed by this change. The primary purpose of this change is to make
the existing baseline ban-honoring functionality easy to maintain.

----

Based on SQUID-800-cache-mgr-caching commit b4e7aad.