Hack: Make Optional<bool> trivially copyable
    ipc/TypedMsgHdr.h: static assertion failed: putPod() used for a POD
    ActionParams.cc:44: required from here [with Pod = RequestFlags]

The known XXX in Optional destructor has started to bite us, because
RequestFlags are serialized using memcpy(3) and are now using an
Optional<bool> flag. It is possible to mimic standard std::optional
implementations, avoiding that XXX, but that requires rather
sophisticated C++ tricks with placement new() and such. Specializing the
whole Optional is a better alternative for this _temporary_ class IMO.

Test sandbox: https://godbolt.org/z/n5658Msd4