Enable C++17 (instead of C++11) support if needed
Switch to Autoconf-archive provided version of AX_CXX_COMPILE_STDCXX_17
instead of our customized (and stale) AX_CXX_COMPILE_STDCXX_11.
Hopefully, we can continue to use the official version of that macro,
but, if we cannot, a customized version can be easily added when we have
the need for it (based on then-current Autoconf-archive macro).

Fail bootstrap.sh when AX_CXX_COMPILE_STDCXX_17 (or equivalent) is
missing in the build environment.