Removed %icap::request_attempts
There were several problems with the removed implementation:

* It did not count the attempts correctly. It would always log 0 or 1
  because each Icap::Xaction creates its own ALE. ICAP retries are
  handled by higher-level classes. The correct implementation should
  probably use the existing Icap::Xaction::attempts counter.

* It counted eCAP attempts but did not provide a logformat %code for
  reporting them when ICAP support was disabled.

* It probably contradicted the latest master approach to counting only
  finished connection attempts.

* Its documentation was very incomplete. I bet most admins would not
  have guessed what ICAP request attempts are (counted).

Given the number of implementation problems and the existing
complexities of http::request_attempts support, as well as no specific
demand for this feature (that I recall), it feels like we should not
rush with adding this ICAP/eCAP feature. We should add it later, after
the primary feature is finalized and merged.