Bug 5254 p1: Do not leak master process' cache.log to kids
The fork()ed kids unknowingly inherit cache.log file descriptor and,
hence, prevent the underlying file from being deleted on log rotation.

This change does not fully fix the bug because the file is still being
held open by the master process itself. This change was isolated because
it lacks bad/controversial side effects -- it is a simple step forward.