Remove broken -sha1 option from server_cert_fingerprint
server_cert_fingerprint support for the sha1 parameter has been broken
for years, probably since its inception (2012 commit 42d3334). The bug
was known since at least 2018 when it was mentioned in Bug 4847
discussion.  The single-dash syntax violates the double-dash pattern
used for other -long ACL options. If fixed, using the option would not
change Squid behavior because SHA1 is the default (and the only
supported) fingerprinting algorithm. The option was meant to
allow admins to be explicit about that default in case it changes in the
future, but implementation bugs derailed that plan. The fix is not
straightforward, and we should be focusing on more important things.