Accumulate disk reads to parse response headers larger than 4096 bytes
This code itself does not depend on the given buffer size, but its
Assure() checks will fail if it cannot parse headers because the given
buffer is full. The current given buffer is 4096 bytes or smaller, so
those checks will fail if they discover a bigger header. To fix this, we
need to modify callers to give us bigger buffers.