Removed a rather misleading/awkward official assertion
... to eliminate one hard-coded case of "all store_client copy() buffers
are HTTP_REQBUF_SZ bytes in size" logic.

The official assert was problematic for many reasons, including:

* The code where assert() is located does not care about the condition.

* The exact meaning of the condition is very difficult to describe
  accurately. I bet the condition "evolved" through trial-and-error as
  ClientStreams were taking over the code. For example, commit c8be6d7
  added the flags.headersSent exception, probably because the earlier
  commit edce4d9 added a seemingly bogus condition that could have only
  worked for the first copy() callback. We now set flags.headersSent
  asynchronously and much later, but the flag was set directly in
  clientSendMoreData() when the exception was added. The flag meaning
  has changed.

* The assertion, if interpreted as if it were written now, is rather
  weak. It kind of tries to check that only allowed HTTP reply bodies
  are received, but it misses the first callback that may have a body as
  well. The code handles that case correctly, but the assertion does not
  and cannot cover it.

I tried to add an Assure(storeReadBuffer.contains(result)) instead, but
we cannot even do that because we do not know for sure what buffer we
are receiving. The "we've got to copy some data" code a few lines
earlier imply that it may not be a storeReadBuffer-derived buffer. If it
is not, then we have no way of checking whether the length value is
legitimate because StoreIOBuffer does not know its buffer capacity!