More renames
This commit applies safer/dumber script changes that require less review
diligence (but break build). The upcoming commit contains a few riskier
adjustments that require a more careful review (and fix build).

The script identified and renamed the following 28 classes:

* testEventLoop
* testHttp1Parser
* testHttpReply
* testHttpRequest
* testHttpRequestMethod
* testIcmp
* testIpAddress
* testLookupTable
* testMem
* testNetDb
* testPackableStream
* testPreCompiler
* testRFC1035
* testRFC1738
* testRefCount
* testRock
* testSBuf
* testSBufList
* testStatHist
* testStore
* testStoreController
* testStoreHashIndex
* testString
* testTokenizer
* testURL
* testUfs
* testUriScheme
* testYesNoNone

Also renamed former TestStore to StoreControllerStub to avoid a clash
with the renamed testStore. Arguably, it is also a better name for that
class.