CI: Do not mention other branches in master CI config
When the original GitHub Actions configuration was written (see commit
2ed8cc4), we incorrectly assumed that GitHub will use that single
configuration (stored in the master branch) for testing all repository
branches. In reality, GitHub uses configuration from the being-tested
branch. Thus, we do not need to (and, hence, should not) list any
specific branches in the master branch configuration, with the exception
of the "master" branch itself and the "auto" branch where all PRs
(including master-targeting PRs) are staged.