Documented two old problems
I suspect that one or both of these problems contribute to
proxy-collapsed-forwarding test failures in pending branch code because
those pending branch changes alter the timing of some events. I know the
failures are happening because Rock swapout is still closing during
these collapsed hits (and Rock does not support swapin during swapout),
and because we trimmed the swapped out header bytes from memory already,
but it is not yet clear why the official code does not suffer from the
same problem.

    attachToDisk: attaching entry ... SWAPOUT_WRITING
    ...
    storeClientType: e:tw3753d5060@0=sw1p2V*3 inmem_lo=401
    store_client.cc(386) doCopy: co: 0+0 hi: 402
    store_client.cc(446) ... Need to open swap in file
    store_swapin.cc(33) storeSwapInStart: e:tw6604d1967@0=sw1p2V/
    store_client.cc(637) fail: once after 0
    ...
    storeSwapOutFileClosed: SwapOut complete