Fix rock/RockSwapDir.cc "slot->sameKey()" assertion
Introduced in 2014 commit fa2301d, the assertion was too strict: Due to
store_client idiosyncrasies, openStoreIO() may be called long after the
initial cache hit is detected and locked for reading. It is possible for
the StoreEntry to go private (and be marked for release) since then.
Private StoreEntry objects lose their original public keys. The existing
anchor lock assures that we still point to the right disk entry, but we
cannot double check using a private key.

There are no known sightings of this assertion in the wild, but we can
now reliably reproduce it using high-load Web Polygraph tests. It is not
clear whether some other Squid changes made the triggering sequence of
events more likely or we were just lucky to discover the right workload.